Further refactored TemporaryImportFile => Also renamed it to EmbeddedResourceFileWriter Related to Issue [WTI-360]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@2539 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 7e9ae065d1be0c9a10bfb2a36c352c2cec8685df
Former-commit-id: 161649113bf5e355f3056ba96e1a20102fc34325

+++ Demo/Ringtoets/src/Demo.Ringtoets/EmbeddedResourceFileWriter.cs
@@ -5,19 +5,22 @@ using Core.Common.Utils.Reflection;
 namespace Demo.Ringtoets
 {
     /// <summary>
-    /// Class for creating a temporary file in the windows Temp directory based on a file 
-    /// stored in Embedded Resources.
+    /// Class for writing Embedded Resources to the Windows Temp directory.
     /// </summary>
-    internal class TemporaryImportFile : IDisposable
+    internal class EmbeddedResourceFileWriter : IDisposable
     {
+        private readonly bool removeFilesOnDispose;
         private readonly string targetFolderPath;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="TemporaryImportFile"/> class.
+        /// Initializes a new instance of the <see cref="EmbeddedResourceFileWriter"/> class.
         /// </summary>
-        /// <param name="embeddedResourceFileNames">The names of the 'Embedded Resource' files to temporary copy to the file system.</param>
-        public TemporaryImportFile(params string[] embeddedResourceFileNames)
+        /// <param name="removeFilesOnDispose">Whether or not the files should be removed after disposing the created <see cref="EmbeddedResourceFileWriter"/> instance.</param>
+        /// <param name="embeddedResourceFileNames">The names of the Embedded Resource files to (temporary) write to the Windows Temp directory.</param>
+        public EmbeddedResourceFileWriter(bool removeFilesOnDispose, params string[] embeddedResourceFileNames)
         {
+            this.removeFilesOnDispose = removeFilesOnDispose;
+
             targetFolderPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
 
             Directory.CreateDirectory(targetFolderPath);
@@ -28,6 +31,9 @@ namespace Demo.Ringtoets
             }
         }
 
+        /// <summary>
+        /// Gets the target folder path.
+        /// </summary>
         public string TargetFolderPath
         {
             get
@@ -36,15 +42,20 @@ namespace Demo.Ringtoets
             }
         }
 
+        /// <summary>
+        /// Disposes the <see cref="EmbeddedResourceFileWriter"/> instance.
+        /// </summary>
         public void Dispose()
         {
-            Directory.Delete(targetFolderPath, true);
+            if (removeFilesOnDispose)
+            {
+                Directory.Delete(targetFolderPath, true);
+            }
         }
 
         private void WriteEmbeddedResourceToTemporaryFile(string embeddedResourceFileName, string filePath)
         {
             var stream = GetStreamToFileInResource(embeddedResourceFileName);
-
             var bytes = GetBinaryDataOfStream(stream);
 
             File.WriteAllBytes(filePath, bytes);
